requirements.txt
absl-py==0.13.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
clang==5.0
cycler==0.10.0
et-xmlfile==1.1.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.2
importlib-metadata==4.11.1
joblib==1.0.1
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libclang==11.1.0
Markdown==3.3.4
matplotlib==3.4.2
natsort==7.1.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.52
openpyxl==3.0.9
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.4.1
SimpleITK==2.1.1
six==1.15.0
sklearn==0.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
tensorflow-gpu-estimator==2.2.0
tensorflow-io-gcs-filesystem==0.20.0
termcolor==1.1.0
threadpoolctl==2.2.0
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.7.0
